Implements --allow-dirty on sync method
This allows you to make post-commits on a repository that will only
effect the working copy, (such as to the vendor-sync file) but will not
effect the tag based workflow.

Its common to add new things to a project that are meta, and to not
update your vendor-rc file (while it is contained in version control).
Whatever the reason, the tool should not encourage a forced update
workflow for exporting.